CB-9241 Create database tables and entities for load balancer
Creates the database tables used to store load balancer and target group
information, where a target group is a set of instance groups traffic will be
forwarded to by the load balancer, and a type that determines what ports are
used for forwarding. Sets up a relationship with the existing Stack and
InstanceGroup entities such that:

Stack --OneToMany--> LoadBalancer --OneToMany--> TargetGroup
TargetGroup <--ManyToMany--> InstanceGroup.

Creates the services used to query the LoadBalancer and TargetGroup tables.