CB-9241 Create load balancer from stack request
When a StackV4Request is received, parse it to determine if a load balancer
needs to be created. If the Knox gateway service is explicitly or implicitly
defined, create a load balancer and target group for a Knox frontend load
balancer. Add this load balancer to the Stack object. When the Stack is
converted to a CloudStack, convert the Stack load balance to a CloadLoadBalancer.
Adds new units tests to test both of these conversions.