CB-9365 Create load balancer as part of AWS CloudFormation stack
Expands the AWS Cloudformation templates to allow for the creation of load
balancers, target groups, and listeners in the Cloudformation stack. These are
created and configured as updates to the CF stack after its initial creation.
Also adds a retry to the CF deletion logic if the intial deletion fails. This
is to work around a problem with CF not deleting the load balancer assets in
the correct order, resulting in failed deletions.